add command to gen docs
