+++ crates/sui-graphql-rpc/src/main.rs
@@ -25,6 +25,17 @@ async fn main() {
                 println!("{}", &out);
             }
         }
+        Command::GenerateExamples { file } => {
+            let new_content: String = sui_graphql_rpc::examples::generate_markdown();
+
+            let mut buf: PathBuf = PathBuf::from(env!("CARGO_MANIFEST_DIR"));
+            buf.push("docs");
+            buf.push("examples.md");
+            let file = file.unwrap_or(buf);
+
+            std::fs::write(file.clone(), &new_content).unwrap();
+            println!("Written examples to file: {:?}", file);
+        }
         Command::StartServer {
             db_url,
             port,
